requirements.txt
# Core dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1

# AI Provider SDKs
anthropic==0.18.1
google-generativeai==0.3.2
openai==1.12.0
pillow==10.2.0  # Required for Google Gemini

# Web app dependencies (for labeling interface)
flask==3.0.0
flask-cors==4.0.0

# Optional but recommended
tqdm==4.66.1  # Progress bars
tenacity==8.2.3  # Advanced retry logic